Add clientside archiving functionality
Override deleteTiddler so that it moves the contents to an archive bag
instead of actually deleting it.

This affects config.adaptors.tiddlyweb.deleteTiddler. The old deleteTiddler
is still available as _deleteTiddler, should the old behaviour be required for
any purpose.

The tiddler is saved in the archive bag as:

TiddlerName [date]

eg: FooBar [30 Nov 2010 17:47:35]

NB: This relies on this (as yet unreleased) commit to TiddlyWebAdaptor (version
1.4.8): http://trac.tiddlywiki.org/changeset/12610