Changed procedure for setting deleteTiddler back again before move
The old deleteTiddler (the original one) is saved as _deleteTiddler.
moveTiddler requires this to be used as deleteTiddler in order for things
not to mess up (it expects deleteTiddler to actually delete, and this is
what we want).

Previously this was being handled by setting deleteTiddler back to
_deleteTiddler inside the new deleteTiddler function. While this was
fine to a point, it means that you can only do one "delete" (ie move
to archive) per adaptor instance.

A safer approach is to create a new instance inside deleteTiddler, set
_deleteTiddler on just that instance, and call moveTiddler with that instance
instead. That is what this commit does.